x86: h264qpel: int --> ptrdiff_t for some line_size parameters

+++ libavcodec/x86/h264_qpel.c
@@ -28,17 +28,20 @@
 #include "dsputil_mmx.h"
 
 #if HAVE_YASM
-void ff_put_pixels4_mmxext(uint8_t *block, const uint8_t *pixels, int line_size, int h);
-void ff_avg_pixels4_mmxext(uint8_t *block, const uint8_t *pixels, int line_size, int h);
-void ff_put_pixels8_mmxext(uint8_t *block, const uint8_t *pixels, int line_size, int h);
+void ff_put_pixels4_mmxext(uint8_t *block, const uint8_t *pixels,
+                           ptrdiff_t line_size, int h);
+void ff_avg_pixels4_mmxext(uint8_t *block, const uint8_t *pixels,
+                           ptrdiff_t line_size, int h);
+void ff_put_pixels8_mmxext(uint8_t *block, const uint8_t *pixels,
+                           ptrdiff_t line_size, int h);
 static void ff_put_pixels16_mmxext(uint8_t *block, const uint8_t *pixels,
-                                   int line_size, int h)
+                                   ptrdiff_t line_size, int h)
 {
     ff_put_pixels8_mmxext(block,     pixels,     line_size, h);
     ff_put_pixels8_mmxext(block + 8, pixels + 8, line_size, h);
 }
 static void ff_avg_pixels16_mmxext(uint8_t *block, const uint8_t *pixels,
-                                   int line_size, int h)
+                                   ptrdiff_t line_size, int h)
 {
     ff_avg_pixels8_mmxext(block,     pixels,     line_size, h);
     ff_avg_pixels8_mmxext(block + 8, pixels + 8, line_size, h);
